Refactor, create floor map renderer helper containing common draw methods
