+++ features/il_display_assets_map/lib/src/widgets/assets_painter.dart
@@ -1,24 +1,27 @@
 import 'package:flutter/material.dart';
 import 'package:il_core/il_entities.dart';
+import 'package:il_core/il_helpers.dart';
 import 'package:il_core/il_theme.dart';
-import 'package:vector_math/vector_math_64.dart' show Vector3;
 
 class AssetsPainter extends CustomPainter {
   final Matrix4 transform;
   final FloorMap floorMap;
   final List<Asset> assets;
 
+  late FloorMapRenderer floorMapRenderer;
+
   AssetsPainter({
     required this.transform,
     required this.floorMap,
     required this.assets,
-  });
+  }) {
+    floorMapRenderer = FloorMapRenderer();
+    floorMapRenderer.transform = transform;
+  }
 
   @override
   void paint(Canvas canvas, Size size) {
-    for (var zone in floorMap.zones!) {
-      drawZoneLabel(zone, canvas);
-    }
+    floorMapRenderer.drawZoneLabels(canvas, floorMap);
 
     for (var asset in assets) {
       drawAsset(asset, canvas);
@@ -33,7 +36,7 @@ class AssetsPainter extends CustomPainter {
     // transform asset position
 
     double scale = transform.getMaxScaleOnAxis();
-    Vector3 translation = transform.getTranslation();
+    var translation = transform.getTranslation();
 
     Rect trackingArea = floorMap.trackingArea;
 
@@ -68,44 +71,8 @@ class AssetsPainter extends CustomPainter {
     var labelRectPosition = labelPosition - const Offset(12, 2);
     var labelRectSize = Offset(textPainter.width + 24, textPainter.height + 4);
 
-    _drawRRect(labelRectPosition, labelRectSize, labelRectSize.dy / 2, canvas, paint);
+    floorMapRenderer.drawRRect(canvas, labelRectPosition, labelRectSize, labelRectSize.dy / 2, paint);
     textPainter.paint(canvas, labelPosition);
-  }
-
-  void drawZoneLabel(FloorMapZone zone, Canvas canvas) {
-    const double labelFontSize = 16;
-
-    // transform zone label position
-
-    double scale = transform.getMaxScaleOnAxis();
-    Vector3 translation = transform.getTranslation();
-
-    Rect trackingArea = floorMap.trackingArea;
-
-    double zx = zone.labelPoint.dx + trackingArea.left;
-    double zy = zone.labelPoint.dy + trackingArea.top;
-
-    double tx = zx * scale + translation.x;
-    double ty = zy * scale + translation.y;
-
-    // draw zone name
-
-    var textPainter = TextPainter(
-      textDirection: TextDirection.ltr,
-      text: TextSpan(
-        text: zone.name,
-        style: const TextStyle(color: Colors.black, fontSize: labelFontSize),
-      ),
-    );
-
-    textPainter.layout();
-    ty -= textPainter.height;
-
-    textPainter.paint(canvas, Offset(tx, ty));
-  }
-
-  void _drawRRect(Offset pos, Offset size, double r, Canvas canvas, Paint paint) {
-    canvas.drawRRect(RRect.fromRectXY(Rect.fromLTWH(pos.dx, pos.dy, size.dx, size.dy), r, r), paint);
   }
 
   @override
