+++ features/il_display_assets_map/pubspec.yaml
@@ -12,7 +12,6 @@ dependencies:
     sdk: flutter
   
   flutter_svg: ^2.0.16
-  vector_math: ^2.1.4
   font_awesome_flutter: ^10.8.0
 
   il_core:
