+++ il_core/lib/il_helpers.dart
@@ -1,3 +1,4 @@
 library helpers;
 
 export 'src/logic/helpers/geometry_helper.dart';
+export 'src/logic/helpers/floor_map_renderer.dart';
