+++ il_core/pubspec.yaml
@@ -12,6 +12,7 @@ dependencies:
     sdk: flutter
   
   intl: ^0.20.1
+  flutter_svg: ^2.0.16
 
 dev_dependencies:
   flutter_test:
